# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kouvee Petshop API Documentation",
        "contact": {
            "name": "Natan Hari P.",
            "email": "natanhp@protonmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customers/getall": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Get all customers",
                "operationId": "App\\Http\\Controllers\\CustomersController::getAll",
                "responses": {
                    "default": {
                        "description": "Get all customers"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/insert": {
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Insert a new customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::insert",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the customer",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "The address of the customer",
                                        "type": "string"
                                    },
                                    "dateBirth": {
                                        "description": "The birth date of the customer",
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "description": "The phone number of the customer",
                                        "type": "string"
                                    },
                                    "createdBy": {
                                        "description": "The foreign key of the cs who creates the customer",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/getbyid/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Get a customer by id",
                "operationId": "App\\Http\\Controllers\\CustomersController::getCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a customer by id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/getbyname/{name}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Get a customer by name",
                "operationId": "App\\Http\\Controllers\\CustomersController::getCustomerByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a customer by name"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/update": {
            "put": {
                "tags": [
                    "customers"
                ],
                "summary": "Update a customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "The id of the customer",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "The name of the customer",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "The address of the customer",
                                        "type": "string"
                                    },
                                    "dateBirth": {
                                        "description": "The birth date of the customer",
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "description": "The phone number of the customer",
                                        "type": "string"
                                    },
                                    "updatedBy": {
                                        "description": "The foreign key of the cs who updates the customer",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/delete/{id}/{csId}": {
            "delete": {
                "tags": [
                    "customers"
                ],
                "summary": "Deletes a customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "csId",
                        "in": "path",
                        "description": "CS who delted the customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Customer not deleted it's because either the deletion failed or customer to be deleted not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/restore/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Restore the delted customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted customer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/getallpets/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Get all of the customer's pets by customer id",
                "operationId": "App\\Http\\Controllers\\CustomersController::getAllCustomerPetsByCustomerId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "All of customer's pets"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/employees/restore/{id}": {
            "get": {
                "tags": [
                    "employees"
                ],
                "description": "Restore the delted employee",
                "operationId": "App\\Http\\Controllers\\EmployeesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted employee"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/petsizes/restore/{id}": {
            "get": {
                "tags": [
                    "pet sizes"
                ],
                "description": "Restore the deleted pet size",
                "operationId": "App\\Http\\Controllers\\PetSizesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet size",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet size"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pettypes/restore/{id}": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Restore the deleted pet type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pets/restore/{id}": {
            "get": {
                "tags": [
                    "pets"
                ],
                "description": "Restore the deleted pet",
                "operationId": "App\\Http\\Controllers\\PetsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/products/restore/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "description": "Restore the delted product",
                "operationId": "App\\Http\\Controllers\\ProductsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted product"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/servicedetails/restore/{id}": {
            "get": {
                "tags": [
                    "service details"
                ],
                "description": "Restore the deleted service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service detail"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/services/restore/{id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "description": "Restore the deleted service",
                "operationId": "App\\Http\\Controllers\\ServicesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/suppliers/restore/{id}": {
            "get": {
                "tags": [
                    "suppliers"
                ],
                "description": "Restore the delted supplier",
                "operationId": "App\\Http\\Controllers\\SuppliersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restored the deleted supplier"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}